\name{BOR1}
\alias{BOR1}
\docType{data}
\title{
Is a vector of Bed Occupation Rates at the beginning of study
}
\description{
Is a vector of Bed Occupation Rates at the beginning of study
}
\usage{data(BOR1)}
\format{
  The format is:
 num [1:16] 72.5 48.9 42.8 40.8 60 ...
}

\examples{
BOR1=c(72.54,48.86,42.77,40.81,60,28.61,20.29,12.84,100,47.07,78.51,45,49,20,88,90)
}
\keyword{datasets}
