\name{BOR2}
\alias{BOR2}
\docType{data}
\title{
Is a vector of Bed Occupation Rates at the end of study
}
\description{
Is a vector of Bed Occupation Rates at the end of study
}
\usage{data(BOR2)}
\format{
  The format is:
 num [1:16] 12.05 12.5 6.83 4.35 5.33 ...
}

\examples{
BOR2=c(12.05,12.5,6.83,4.35,5.33,7.77,6.28,2.73,35.07,13.23,12.21,12,4,5,25,16)
}
\keyword{datasets}
