\name{BTR1}
\alias{BTR1}
\docType{data}
\title{
Is a vector of Bed TurnOver Rates at the beginning of study
}
\description{
Is a vector of Bed TurnOver Rates at the beginning of study
}
\usage{data(BTR1)}
\format{
  The format is:
 num [1:16] 40 55.37 80 17.33 40 ...
}

\examples{
BOR1=c(40,55.37,80,17.33,40,84,57.91,9.73,30,34.09,57.52,50,55,30,35,20)
}
\keyword{datasets}
