\name{BTR2}
\alias{BTR2}
\docType{data}
\title{
Is a vector of Bed TurnOver Rates at the end of study
}
\description{
Is a vector of Bed TurnOver Rates at the end of study
}
\usage{data(BTR2)}
\format{
  The format is:
 num [1:16] 15.22 15.58 6.52 2.35 10.75 ...
}

\examples{
BOR1=c(15.22,15.58,6.52,2.35,10.75,25,27.11,1.867,9,9.695,11.20,5,14,25,23,26)
}
\keyword{datasets}
